Reverts MakeHardlinks loop to use structured bindings
I checked that the bindings can work via a simple unit test which was thrown out after verification.
It seems the json library throws an exception if the identifier is asked to convert into the wrong type.
Subsequently it seems that the type the IDE claims is false and it is actually more complicated than the structured bindings going straight to the types needed.